Fixed typehints and added stubs for missing packages. For some reason we the ci was missconfigured, should be fixed now.
