+++ .github/workflows/python.yml
@@ -27,11 +27,13 @@ jobs:
                   pip install .[typed]
             - name: Check style with Ruff
               continue-on-error: true
+              id: ruff
               run: |
                   cd ./backend
                   ruff check --output-format=github .
             - name: Check type hints with mypy
               continue-on-error: true
+              id: mypy
               run: |
                   cd ./backend
                   mypy --show-error-codes --check-untyped-defs --config-file ./pyproject.toml ./beets_flask
