+++ backend/beets_flask/database/models/states.py
@@ -18,7 +18,6 @@ import pickle
 from pathlib import Path
 from typing import List, Optional
 
-import sqlalchemy.types as types
 from beets.importer import ImportTask, action, library
 from sqlalchemy import (
     ForeignKey,
@@ -456,32 +455,6 @@ class TaskStateInDb(Base):
         return self.to_live_state().serialize()
 
 
-class IntDictType(types.TypeDecorator):
-    """
-    Stores a dict[int, int] as a JSON-encoded string in the database.
-    """
-
-    impl = types.Text
-    cache_ok = True
-
-    def process_bind_param(self, value, dialect):
-        if value is None:
-            return None
-        if not isinstance(value, dict) or not all(
-            isinstance(k, int) and isinstance(v, int) for k, v in value.items()
-        ):
-            raise ValueError("Value must be a dict[int, int]")
-        return json.dumps({str(k): v for k, v in value.items()})
-
-    def process_result_value(self, value, dialect):
-        if value is None:
-            return None
-        return {int(k): v for k, v in json.loads(value).items()}
-
-    def copy(self, **kw):
-        return IntDictType(self.impl.length)  # type: ignore
-
-
 class CandidateStateInDb(Base):
     """Represents a candidate (potential match) for an import task.
 
@@ -504,9 +477,7 @@ class CandidateStateInDb(Base):
     duplicate_ids: Mapped[str]
 
     # association between tracks online and items on disk, from int to int
-    mapping: Mapped[IntDictType] = mapped_column(
-        IntDictType, default=None, nullable=True
-    )
+    mapping: Mapped[dict[int, int] | None]
 
     def __init__(
         self,
@@ -538,7 +509,6 @@ class CandidateStateInDb(Base):
             match=state.match,
             duplicate_ids=state.duplicate_ids,
         )
-        # hmm not sure why we get the type error
         candidate.mapping = state._mapping
         return candidate
